src/maintenanceService/requirements.txt
alembic==1.17.1
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
PyMySQL==1.1.2
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
